# web/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Clipaso API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/plans": {
      "get": {
        "tags": [
          "account"
        ],
        "summary": "List Plans",
        "description": "Planes públicos (página de precios).",
        "operationId": "list_plans_plans_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/PlanOut"
                  },
                  "type": "array",
                  "title": "Response List Plans Plans Get"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/me": {
      "get": {
        "tags": [
          "account"
        ],
        "summary": "Me",
        "operationId": "me_me_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeOut"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "account"
        ],
        "summary": "Delete Me",
        "description": "Elimina la cuenta: vídeos, clips, historial y consumo, y después el usuario de Supabase Auth.",
        "operationId": "delete_me_me_delete",
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/options": {
      "get": {
        "tags": [
          "account"
        ],
        "summary": "Clip Options",
        "description": "Formatos, duraciones y estilos de subtítulos disponibles (los pinta la web).",
        "operationId": "clip_options_options_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OptionsOut"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/me/preferences": {
      "get": {
        "tags": [
          "account"
        ],
        "summary": "Get Preferences",
        "operationId": "get_preferences_me_preferences_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PreferencesOut"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "account"
        ],
        "summary": "Put Preferences",
        "description": "Marca personal para los próximos vídeos (los estilos de subtítulos van en /me/styles).",
        "operationId": "put_preferences_me_preferences_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PreferencesIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PreferencesOut"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/me/logo": {
      "put": {
        "tags": [
          "account"
        ],
        "summary": "Put Logo",
        "description": "Sube el logo (PNG o JPG, máx. 1 MB) como cuerpo de la petición.",
        "operationId": "put_logo_me_logo_put",
        "requestBody": {
          "content": {
            "image/png": {
              "schema": {
                "type": "string",
                "format": "binary"
              }
            },
            "image/jpeg": {
              "schema": {
                "type": "string",
                "format": "binary"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PreferencesOut"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "account"
        ],
        "summary": "Remove Logo",
        "operationId": "remove_logo_me_logo_delete",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PreferencesOut"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/me/locale": {
      "put": {
        "tags": [
          "account"
        ],
        "summary": "Put Locale",
        "description": "Idioma de la cuenta: el de la web y el de los emails (clips listos, fallos).",
        "operationId": "put_locale_me_locale_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LocaleIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/me/styles": {
      "get": {
        "tags": [
          "account"
        ],
        "summary": "List Styles",
        "description": "Estilos de serie (con los cambios del usuario) y estilos propios.",
        "operationId": "list_styles_me_styles_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StylesOut"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "account"
        ],
        "summary": "Create Style",
        "operationId": "create_style_me_styles_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StyleIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StylesOut"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/me/styles/{style_id}": {
      "put": {
        "tags": [
          "account"
        ],
        "summary": "Update Style",
        "description": "Modifica un estilo propio o guarda tu versión de uno de serie.",
        "operationId": "update_style_me_styles__style_id__put",
        "parameters": [
          {
            "name": "style_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Style Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StyleUpdateIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StylesOut"
                }
              }
            }
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "account"
        ],
        "summary": "Delete Style",
        "description": "Borra un estilo propio o devuelve uno de serie a su versión original.",
        "operationId": "delete_style_me_styles__style_id__delete",
        "parameters": [
          {
            "name": "style_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Style Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StylesOut"
                }
              }
            }
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/me/default-style": {
      "put": {
        "tags": [
          "account"
        ],
        "summary": "Set Default Style",
        "operationId": "set_default_style_me_default_style_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DefaultStyleIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StylesOut"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DefaultStyleIn": {
        "properties": {
          "id": {
            "type": "string",
            "maxLength": 40,
            "title": "Id"
          }
        },
        "type": "object",
        "required": [
          "id"
        ],
        "title": "DefaultStyleIn"
      },
...
      "ErrorResponse": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBody"
          }
        },
        "type": "object",
        "required": [
          "error"
        ],
        "title": "ErrorResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "LocaleIn": {
        "properties": {
          "locale": {
            "type": "string",
            "enum": [
              "es",
              "en"
            ],
            "title": "Locale"
          }
        },
        "type": "object",
        "required": [
          "locale"
        ],
        "title": "LocaleIn"
      },
      "MeOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "email": {
            "type": "string",
            "title": "Email"
          },
          "locale": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "es",
                  "en"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Locale",
            "description": "Idioma elegido; null si aún no se ha fijado."
          },
          "plan": {
            "$ref": "#/components/schemas/PlanOut"
          },
          "usage": {
            "$ref": "#/components/schemas/UsageOut"
          }
        },
        "type": "object",
        "required": [
          "id",
          "email",
          "locale",
          "plan",
          "usage"
        ],
        "title": "MeOut"
      },
...
      "OptionsOut": {
        "properties": {
          "formats": {
            "items": {
              "$ref": "#/components/schemas/OptionItem"
            },
            "type": "array",
            "title": "Formats"
          },
          "durations": {
            "items": {
              "$ref": "#/components/schemas/OptionItem"
            },
            "type": "array",
            "title": "Durations"
          },
          "presets": {
            "items": {
              "$ref": "#/components/schemas/CaptionPreset"
            },
            "type": "array",
            "title": "Presets"
          },
          "fonts": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Fonts"
          }
        },
        "type": "object",
        "required": [
          "formats",
          "durations",
          "presets",
          "fonts"
        ],
        "title": "OptionsOut"
      },
...
      "PlanOut": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "price_eur_cents": {
            "type": "integer",
            "title": "Price Eur Cents"
          },
          "monthly_minutes": {
            "type": "integer",
            "title": "Monthly Minutes"
          },
          "max_video_minutes": {
            "type": "integer",
            "title": "Max Video Minutes"
          },
          "max_clips_per_job": {
            "type": "integer",
            "title": "Max Clips Per Job"
          },
          "max_concurrent_jobs": {
            "type": "integer",
            "title": "Max Concurrent Jobs"
          },
          "max_upload_mb": {
            "type": "integer",
            "title": "Max Upload Mb"
          },
          "retention_days": {
            "type": "integer",
            "title": "Retention Days"
          }
        },
        "type": "object",
        "required": [
          "code",
          "name",
          "price_eur_cents",
          "monthly_minutes",
          "max_video_minutes",
          "max_clips_per_job",
          "max_concurrent_jobs",
          "max_upload_mb",
          "retention_days"
        ],
        "title": "PlanOut"
      },
      "PreferencesIn": {
        "properties": {
          "branding": {
            "$ref": "#/components/schemas/BrandingPrefs"
          }
        },
        "type": "object",
        "required": [
          "branding"
        ],
        "title": "PreferencesIn"
      },
      "PreferencesOut": {
        "properties": {
          "caption_style": {
            "$ref": "#/components/schemas/CaptionStyle",
            "description": "El estilo por defecto (ver /me/styles)."
          },
          "branding": {
            "$ref": "#/components/schemas/BrandingPrefs"
          },
          "logo_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Logo Url"
          }
        },
        "type": "object",
        "required": [
          "caption_style",
          "branding",
          "logo_url"
        ],
        "title": "PreferencesOut"
      },
...
      "StyleIn": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 40,
            "title": "Name"
          },
          "style": {
            "$ref": "#/components/schemas/CaptionStyle"
          }
        },
        "type": "object",
        "required": [
          "name",
          "style"
        ],
        "title": "StyleIn"
      },
...
      "StylesOut": {
        "properties": {
          "styles": {
            "items": {
              "$ref": "#/components/schemas/UserStyle"
            },
            "type": "array",
            "title": "Styles"
          },
          "default_id": {
            "type": "string",
            "title": "Default Id"
          },
          "max_custom": {
            "type": "integer",
            "title": "Max Custom"
          }
        },
        "type": "object",
        "required": [
          "styles",
          "default_id",
          "max_custom"
        ],
        "title": "StylesOut"
      },
...
    }
  }
}